pulls out job mapping table component into reusable component
